Refactor NuGet.Protocol v3 search: improve method names (#3426)
* Refactor the code after Nikolche code review.

* Put back old constructor along with new constructor for PackageSearchResourceV3.

* Add back Obsolote attributes at methods in Resources/RawSearchResourceV3.

* Address code review comments.

* Make public constructor for Protocol/Resources/PackageSearchResourceV3 to internal.

* Address additional code review.
