+++ src/NuGet.Core/NuGet.Protocol/GlobalSuppressions.cs
@@ -300,4 +300,4 @@ using System.Diagnostics.CodeAnalysis;
 [assembly: SuppressMessage("Build", "CA1001:Type 'RemoteV3FindPackageByIdResource' owns disposable field(s) '_dependencyInfoSemaphore, _httpSource' but is not disposable", Justification = "<Pending>", Scope = "type", Target = "~T:NuGet.Protocol.RemoteV3FindPackageByIdResource")]
 [assembly: SuppressMessage("Build", "CA1001:Type 'ServiceIndexResourceV3Provider' owns disposable field(s) '_semaphore' but is not disposable", Justification = "<Pending>", Scope = "type", Target = "~T:NuGet.Protocol.ServiceIndexResourceV3Provider")]
 [assembly: SuppressMessage("Build", "CA1052:Type 'V2FeedUtilities' is a static holder type but is neither static nor NotInheritable", Justification = "<Pending>", Scope = "type", Target = "~T:NuGet.Protocol.V2FeedUtilities")]
-[assembly: SuppressMessage("Globalization", "CA1308:Normalize strings to uppercase", Justification = "We intentionally making to lower case because we're creating web api query.", Scope = "member", Target = "~M:NuGet.Protocol.RawSearchResourceV3.SearchPageNew``1(System.Func{System.Uri,System.Threading.Tasks.Task{``0}},System.String,NuGet.Protocol.Core.Types.SearchFilter,System.Int32,System.Int32,NuGet.Common.ILogger,System.Threading.CancellationToken)~System.Threading.Tasks.Task{``0}")]
+[assembly: SuppressMessage("Globalization", "CA1308:Normalize strings to uppercase", Justification = "We intentionally making to lower case because we're creating web api query.", Scope = "member", Target = "~M:NuGet.Protocol.PackageSearchResourceV3.SearchPage``1(System.Func{System.Uri,System.Threading.Tasks.Task{``0}},System.String,NuGet.Protocol.Core.Types.SearchFilter,System.Int32,System.Int32,NuGet.Common.ILogger,System.Threading.CancellationToken)~System.Threading.Tasks.Task{``0}")]
