+++ src/NuGet.Core/NuGet.Protocol/Providers/RawSearchResourceV3Provider.cs
@@ -1,14 +1,14 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using NuGet.Protocol.Core.Types;
 
 namespace NuGet.Protocol
 {
+    [Obsolete("Use PackageSearchResource instead (via SourceRepository.GetResourceAsync<PackageSearchResource>")]
     public class RawSearchResourceV3Provider : ResourceProvider
     {
         public RawSearchResourceV3Provider()
