+++ test/NuGet.Core.Tests/NuGet.Protocol.Tests/RawSearchResourceTests.cs
@@ -31,7 +31,9 @@ namespace NuGet.Protocol.Tests
 
             var httpSource = new TestHttpSource(new PackageSource(serviceAddress), responses);
 
+#pragma warning disable CS0618
             var searchResource = new RawSearchResourceV3(httpSource, new Uri[] { new Uri(serviceAddress) });
+#pragma warning restore CS0618
 
             var searchFilter = new SearchFilter(includePrerelease: false)
             {
@@ -42,8 +44,10 @@ namespace NuGet.Protocol.Tests
             var take = 1;
 
             // Act
-            var packages = await searchResource.SearchNew(
-                        "azure b",
+#pragma warning disable CS0618
+            var packages = await searchResource.Search(
+#pragma warning restore CS0618
+                "azure b",
                         searchFilter,
                         skip,
                         take,
@@ -75,7 +79,9 @@ namespace NuGet.Protocol.Tests
             // throw if sync .Read is used
             httpSource.StreamWrapper = (stream) => new NoSyncReadStream(stream);
 
+#pragma warning disable CS0618
             var searchResource = new RawSearchResourceV3(httpSource, new Uri[] { new Uri(serviceAddress) });
+#pragma warning restore CS0618
 
             var searchFilter = new SearchFilter(includePrerelease: false)
             {
@@ -85,8 +91,10 @@ namespace NuGet.Protocol.Tests
             var take = 1;
 
             // Act
-            var packages = await searchResource.SearchNew(
-                        "azure b",
+#pragma warning disable 618
+            var packages = await searchResource.Search(
+#pragma warning restore 618
+                "azure b",
                         searchFilter,
                         skip,
                         take,
@@ -117,7 +125,9 @@ namespace NuGet.Protocol.Tests
 
             httpSource.StreamWrapper = (stream) => new NoSyncReadStream(stream);
 
+#pragma warning disable 618
             var searchResource = new RawSearchResourceV3(httpSource, new Uri[] { new Uri(serviceAddress) });
+#pragma warning restore 618
 
             var searchFilter = new SearchFilter(includePrerelease: false)
             {
@@ -131,8 +141,10 @@ namespace NuGet.Protocol.Tests
 
             // Act/Assert
             await Assert.ThrowsAsync<TaskCanceledException>(() =>
-               searchResource.SearchNew(
-                        "Sentry",
+#pragma warning disable 618
+                searchResource.Search(
+#pragma warning restore 618
+                    "Sentry",
                         searchFilter,
                         skip,
                         take,
